+++ src/onegov.activity/onegov/activity/collections/__init__.py
@@ -1,4 +1,5 @@
 from onegov.activity.collections.activity import ActivityCollection
 from onegov.activity.collections.booking import BookingCollection
+from onegov.activity.collections.occasion import OccasionCollection
 
-__all__ = ['ActivityCollection', 'BookingCollection']
+__all__ = ['ActivityCollection', 'BookingCollection', 'OccasionCollection']