+++ src/onegov.activity/onegov/activity/collections/booking.py
@@ -1,16 +1,12 @@
 from onegov.activity.models import Booking
+from onegov.core.collection import GenericCollection
 
 
-class BookingCollection(object):
+class BookingCollection(GenericCollection):
 
-    def __init__(self, session):
-        self.session = session
-
-    def query(self):
-        return self.session.query(Booking)
-
-    def by_id(self, id):
-        return self.query().filter(Booking.id == id).first()
+    @property
+    def model_class(self):
+        return Booking
 
     def by_user(self, user):
         return self.query().filter(Booking.username == user.username)
@@ -21,19 +17,10 @@ class BookingCollection(object):
     def add(self, user, priority=None, group_code=None,
             last_name=None, first_name=None):
 
-        booking = Booking(
+        return super().add(
             username=user.username,
             priority=priority,
             group_code=group_code,
             last_name=last_name,
             first_name=first_name
         )
-
-        self.session.add(booking)
-        self.session.flush()
-
-        return booking
-
-    def delete(self, booking):
-        self.session.delete(booking)
-        self.sesison.flush()